perf(platform): consolidate build config verification
