+++ turbo/apps/platform/package.json
@@ -13,11 +13,11 @@
     "i18n:check": "i18next-cli extract --ci --quiet && i18next-cli status",
     "generate:emoji-data": "node scripts/generate-chat-thread-emoji-data.node.mjs",
     "lint:emoji-data": "node scripts/generate-chat-thread-emoji-data.node.mjs --check",
-    "lint:build-config": "node --test scripts/check-production-minification.node.mjs && tsx --test scripts/check-single-bundle.node.ts",
+    "lint:build-config": "tsx --test scripts/check-build-config.node.ts",
     "lint:localized-ui": "node --test scripts/check-localized-ui.node.mjs && node scripts/check-localized-ui.mjs",
     "lint:static-assets": "node scripts/check-static-assets.mjs",
     "lint": "pnpm lint:static-assets && pnpm lint:emoji-data && pnpm lint:localized-ui && pnpm lint:build-config && pnpm i18n:check && oxlint --deny-warnings && oxlint --type-aware -c ../../.oxlintrc.typeaware.json --deny-warnings && eslint . --max-warnings 0",
-    "check-types": "tsc -p tsconfig.production.json --noEmit && tsc -p tsconfig.tests.json --noEmit"
+    "check-types": "tsc -p tsconfig.production.json --noEmit && tsc -p tsconfig.tests.json --noEmit && tsc -p tsconfig.build-config.json --noEmit"
   },
   "dependencies": {
     "@clerk/localizations": "^4.13.8",
