+++ turbo/apps/platform/tsconfig.build-config.json
@@ -1,0 +1,4 @@
+{
+  "extends": "./tsconfig.json",
+  "include": ["scripts/check-build-config.node.ts"]
+}
