+++ turbo/knip.json
@@ -35,7 +35,7 @@
         "src/main.ts",
         "i18next.config.ts",
         "scripts/bench-p90-reporter.ts",
-        "scripts/check-single-bundle.node.ts",
+        "scripts/check-build-config.node.ts",
         "src/**/__benches__/**/*.bench.tsx",
         "**/*.{test,spec}.ts?(x)"
       ],
